fix(clerk-js): Handle gracefully Coinbase Wallet initial configuration
There is a chance that as a first time visitor when you try to setup and use the
Coinbase Wallet from scratch in order to authenticate, the initial generate signature
request to be rejected. For this reason we retry the request once more in order for the
flow to be able to be completed successfully.