fix(backend): Handle "Cannot convert argument to a ByteString" errors
Headers.set can throw a "Cannot convert argument to a ByteString" when trying to set/append a string containing unicode characters as headers can only accept ASCII chars in the range of [0-255]. Under certain cases, verifyToken can throw an error containing the unicode character that caused the error which then we would try to include in the debug header.

This would cause the *only current request* to fail so we are adding extra error handling to prevent this edge scenario.