fix(clerk-js): Handle gracefully Coinbase Wallet use of existing Passkey
There is a chance that as a user when you try to setup and use the Coinbase Wallet with
an existing Passkey in order to authenticate, the initial generate signature request to
be rejected. For this reason we retry the request once more in order for the flow to be
able to be completed successfully.