chore: Introduce Passkeys library for expo
In this commit, we are introducing a library to handle passkeys for Expo and Clerk. The library manages the creation and retrieval of passkeys for iOS, Android devices, and Expo web.

We have also provided an example app that serves as a playground and helps us test the clerk-expo libraries. The README includes instructions on how to set up the app for iOS devices, and we will soon provide instructions for Android devices as well.