Update nl-NL.ts
The translation was wrong. In english the text is 'Already have access?' and 'Sign in'. And when clicking you are indeed taking to the login screen.
In dutch the question was 'Don't have an account yet? (Nog geen account?)' and 'Neem deel aan de wachtlijst (Join the waiting list)' which then takes you to a login screen.
Now it's properly translated.