Update `setActive` to handle in-component tasks navigation
The following changes perform some conditional changes on `setActive` logic based on whether the session has new tasks to resolve.

In order to perform in-component navigation for the session task, it must have access to the React context routing state and also not set the transitive state otherwise it'll unmount SignIn/SignUp components.

The `navigate` function will be passed on the `setActive` calls within the React components.

It must also block any redirections to `afterSignInUrl` or `afterSignUpUrl` that are passed as `redirectUrl`