Do not `#emit` on client piggybacking transition to pending
On `updateClient`, if the new client response session is transition from `pending`, but the current session is `active`, we want to avoid emitting the state changes

Otherwise, components could re-render with unexpected UI states. For a custom organization switcher, it'd have a undefined `organization` and render a 'Personal Account', which looks weird even behind the modal backdrop.