Do not pass `redirectUrlComplete` to `navigateToTask`
`handleRedirectCallback` calls `setActive` with `redirectUrl` - this gets used to navigate to once the session transitions to `active` status.

For custom flows, we don't recommend using `redirectUrl`, neither mention in our docs. This commit avoids passing `redirectUrlComplete` to `navigateToTask` to prevent race conditions on the navigation if the session is `active`.