Remove redirect to task within `withRedirect` guard
`withAfterToSignIn` and `withAfterToSignUp` guards were redirecting to a tasks URL with a warning: "Cannot render SignIn/SignUp if there's an existing task"

This is wrong, cause SignIn/SignUp are rendering, but they should go to their initial route as tasks instead. If you try to access `/sign-in`, then the component should navigate to `/tasks`

A follow-up commit will introducer a proper fix.