fix(clerk-js): Show password in Security tab if password is enabled
The Clerk Dashboard now has improved support for enabling an instance
configuration with password enabled, but not required. Such a
configuration makes passwords optional on sign-up, allows users to add
them to accounts, and allows users to use them for sign-in. If the
instance is configured with password enabled but not required, we need
to still show the password in the user's Security tab so they can set or
update their password.

This is implemented by adding a
`UserSettings.instanceIsPasswordEnabled()` property which is true if
password is enabled on an instance, and using this to determine if we
should show the password card on the Security tab.