fix(clerk-js): If password is enabled, instanceIsPasswordBased is true
The Clerk Dashboard now has improved support for enabling an instance
configuration with password enabled, but not required. Such a
configuration makes passwords optional on sign-up, allows users to add
them to accounts, and allows users to use them for sign-in. Previously,
passwords were both enabled and required together.

If the instance is configured with password enabled but not required,
the SDK should still treat the instance as password based. Modify
`UserSettings.instanceIsPasswordBased` to return true in this
configuration. This (a) makes sure that the password card is displayed
on the SecurityPage so that users can modify their passwords, and (b)
allows instant password sign in.

Fixes USER-3245. Tested with unit testing and with local Clerk.

Revised and simplified version of
https://github.com/clerk/javascript/pull/6592 which modifies
`UserSettings.instanceIsPasswordBased` instead of adding a new property.